requirements.txt
django
django-mptt
django-polymorphic
django-polymorphic_tree
requests
paho-mqtt
json
PyMySQL
djangorestframework
virtualenv
pytz